requirements.txt
torch
torchvision
pyyaml
opencv-python
